feat(compartment-mapper): more renames returned from captureFromMap / makeArchiveCompartmentMap
This change isolates the duplicated code in `archive-lite.js` and `capture-lite.js` into a new internal module, `digest.js`.

It provides `digestCompartmentMap()`, which is now consumed by both `archive-lite.js` and `capture-lite.js`.

Both `makeArchiveCompartmentMap()` and `captureFromMap()` now pass new properties from `digestCompartmentMap()` through to their consumers: `newToOldCompartmentMap` and `oldToNewCompartmentMap`.

**The `compartmentRenames` property is now soft-deprecated** (since its naming is ambiguous). The same property is now `newToOldCompartmentMap`; `compartmentRenames` should eventually be removed.

_Note_: The duplicated code drifted slightly in `archive-lite.js`; see e3b310d. These changes were kept and put into `digest.js`.
